% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_value_tree.R
\name{create_collapsible_tree}
\alias{create_collapsible_tree}
\title{Create the collapsible tree diagram}
\usage{
create_collapsible_tree(data, width, height, fontSize)
}
\arguments{
\item{data}{The prepared outcome data}

\item{width}{The width of the tree diagram}

\item{height}{The height of the tree diagram}

\item{fontSize}{The font size for the tree labels}
}
\value{
A collapsible tree diagram
}
\description{
Create the collapsible tree diagram
}
