% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{decision_table}
\alias{decision_table}
\title{Example decision_table used for decision table}
\format{
\subsection{\code{decision_table}}{

A data frame with 48 rows and 5 columns:
\describe{
\item{Type of Outcome}{Type of Outcome}
\item{Summary Statistics for Primary Outcome}{Summary Statistics for Primary Outcome}
\item{Summary Statistics for Treatment Outcome}{Summary Statistics for Treatment Outcome}
\item{Method for Populating CI}{Method for Populating CI}
\item{Required Columns}{Required Columns}

}
}
}
\usage{
decision_table
}
\description{
Example decision_table used for decision table
}
\keyword{datasets}
