% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_chart.R
\name{line_chart}
\alias{line_chart}
\title{Line Chart}
\usage{
line_chart(
  func,
  data_bands,
  data_lines,
  show_bands = "Y",
  show_points = "N",
  line_size = 0.6,
  line_type = "solid",
  xmin,
  xmax,
  ymin,
  ymax,
  xbreaks,
  ybreaks,
  xlab,
  ylab,
  legend_title
)
}
\arguments{
\item{func}{the name of a function, or a call or an expression written as a function of x}

\item{data_bands}{(\code{data.frame}) a data frame with 3 variables named the following:
\enumerate{
\item level: level each band represents
\item ystart: starting point of each band
\item yend: ending point of each band
\item col: color of each band
}}

\item{data_lines}{(\code{data.frame}) a data frame with 7 variables named the following:
\enumerate{
\item stage: stage each line represents
\item xstart: x-axis starting point of each line
\item xend: x-axis ending point of each line
\item y: y-axis position of each line
\item col: color of each line
\item xpos: x-axis position of each annotation (relative to xstart)
\item ypos: y-axis position of each annotation (relative to y)
}}

\item{show_bands}{(\code{character}) a logical value indicating whether to show bands
in the plot (Y, N)}

\item{show_points}{(\code{character}) a logical value indicating whether to show points
at the start of the arrows (Y, N)}

\item{line_size}{(\code{numeric}) size of the line}

\item{line_type}{(\code{character}) type of the line}

\item{xmin}{(\code{numeric}) starting point of x axis}

\item{xmax}{(\code{numeric}) ending point of x axis}

\item{ymin}{(\code{numeric}) starting point of y axis}

\item{ymax}{(\code{numeric}) ending point of y axis}

\item{xbreaks}{(\code{numeric}) x axis breaks}

\item{ybreaks}{(\code{numeric}) y axis breaks}

\item{xlab}{(\code{character}) x axis label}

\item{ylab}{(\code{character}) y axis label}

\item{legend_title}{(\code{character}) legend title}
}
\value{
a ggplot object
}
\description{
Line Chart
}
\examples{
func <- function(x) 47.982 - 0.0305 * x - 0.57525 * x^2 + 0.0245 * x^3

line_chart(func = func,
           data_bands = data_bands,
           data_lines = data_lines,
           xmin = 0,
           xmax = 14.2,
           ymin = 0,
           ymax = 50,
           xbreaks = seq(0, 14, 2),
           ybreaks = seq(0, 50, 5),
           xlab = "Years Since Onset",
           ylab = "Functional Score",
           legend_title = "Severity: "
)
}
