% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_value_tree.R
\name{prepare_binary_data}
\alias{prepare_binary_data}
\title{Prepare binary outcome data}
\usage{
prepare_binary_data(data)
}
\arguments{
\item{data}{The preprocessed data frame}
}
\value{
A data frame with prepared binary outcome data
}
\description{
Prepare binary outcome data
}
