% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_value_tree.R
\name{prepare_continuous_data}
\alias{prepare_continuous_data}
\title{Prepare continuous outcome data}
\usage{
prepare_continuous_data(data)
}
\arguments{
\item{data}{The preprocessed data frame}
}
\value{
A data frame with prepared continuous outcome data
}
\description{
Prepare continuous outcome data
}
