% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_value_tree.R
\name{prepare_outcome_data}
\alias{prepare_outcome_data}
\title{Prepare outcome data for the tree diagram}
\usage{
prepare_outcome_data(data)
}
\arguments{
\item{data}{The preprocessed data frame}
}
\value{
A data frame with prepared outcome data
}
\description{
Prepare outcome data for the tree diagram
}
