% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_value_tree.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess the input data}
\usage{
preprocess_data(data)
}
\arguments{
\item{data}{The input data frame}
}
\value{
A preprocessed data frame
}
\description{
Preprocess the input data
}
