% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid_chart.R
\name{pyramid_chart}
\alias{pyramid_chart}
\title{Pyramid Chart}
\usage{
pyramid_chart(
  data,
  xvar,
  yvar,
  levelvar,
  groupvar,
  xlab,
  alpha_set,
  chartcolors
)
}
\arguments{
\item{data}{\code{dataframe} Demography data containing at least the following columns:
\itemize{
\item \code{Age}: Age groups (factor or character).
\item \code{Gender}: Gender categories (factor or character with two levels).
\item \code{Prevalence}: Numeric values representing the prevalence or count for each age group and gender.
\item \code{Type}: An additional grouping variable (factor or character) to create separate pyramids for different types.
}}

\item{xvar}{\code{value} The column name in \code{data} representing the numeric values to be plotted
on the x-axis (e.g., "Prevalence" or a transformed version of it).}

\item{yvar}{\code{value} The column name in \code{data} representing the age groups to be plotted
on the y-axis (e.g., "Age").}

\item{levelvar}{\code{Factor} The column name in \code{data} representing the grouping variable
to create separate pyramids (e.g., "Type").
Must be a factor, character, or numeric vector.}

\item{groupvar}{\code{Factor} The column name in \code{data} representing the binary groups
to be displayed on each side of the pyramid (e.g., "Gender").
Must be a factor or character with two levels.}

\item{xlab}{\code{text} Label for the x-axis.}

\item{alpha_set}{\code{Value} Specify the transparency of the bars in the chart (between 0 and 1).}

\item{chartcolors}{\code{vector} A vector of two colors to be used for the two sides of the pyramid.}
}
\value{
A ggplot object representing the population pyramid chart.
}
\description{
Creates a population pyramid chart to visualize the distribution of a demographic
variable (e.g., prevalence) across age groups and genders (or other binary groups).
}
\examples{
# Assuming 'demography' is your data frame
demography |>
  dplyr::mutate(
    Type = as.factor(paste0("Type ", Type)), # Create a factor for separate pyramids
    figprev = ifelse(
      Gender == "Females", -1 * Prevalence / 100000, Prevalence / 100000
    ), # Transform prevalence for females to negative values
    Sex = Gender # Rename "Gender" to "Sex" for the plot
  ) |>
  pyramid_chart(
    levelvar = "Type", xvar = "figprev", yvar = "Age",
    groupvar = "Sex", alpha_set = 0.7, chartcolors = colfun()$fig2_colors,
    xlab = "Prevalence (x 100 000)"
  )
}
