% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_tree.R
\name{value_tree}
\alias{value_tree}
\title{Create Value Tree}
\usage{
value_tree(diagram, ...)
}
\arguments{
\item{diagram}{A character string containing the mermaid graph syntax for the value tree.}

\item{...}{Other arguments and parameters you would like to send to
JavaScript.}
}
\value{
A DiagrammeR mermaid object representing the value tree
}
\description{
Generate example value tree using provided mermaid graph syntax.
}
\details{
The function takes a mermaid graph syntax as input to create the value tree.
This allows for flexibility in defining the tree structure and node relationships.

The example provided demonstrates a typical structure based on information from the \code{effects_table}:
\itemize{
\item Top-level node represents the overall Benefit-Risk Balance
\item Second-level nodes represent \code{Factor} for either Benefits or Risks (e.g., Benefit, Risk)
\item Third-level nodes represent \code{Outcomes} for the specific categories of Benefits and Risks
\item Fourth-level nodes include \code{Statistics} for specific endpoints or measures for each Outcome
}

Colors are used to differentiate between different levels or types of effects:
\itemize{
\item Green (#7ABD7E) for top-level and major category nodes
\item Yellow (#FFE733) for secondary category nodes
\item Orange (#FFAA1C) for specific endpoint nodes
\item Grey (#C6C6C6) for nodes which have been pruned from the tree
}

The mermaid syntax used in this function follows these conventions:
\itemize{
\item The graph is defined as left-to-right (LR) using \verb{graph LR;}
\item Nodes are defined with letters (A, B, C, etc.) and their content is enclosed in parentheses
\item Node content is wrapped in \if{html}{\out{<B>}} tags for bold text
\item Relationships between nodes are defined using arrows (-->)
\item Node styles, including fill colors, are defined using the \code{style} keyword
}

For example:
A(\if{html}{\out{<B>}}Node A\if{html}{\out{</B>}})-->B(\if{html}{\out{<B>}}Node B\if{html}{\out{</B>}})
style A fill:#7ABD7E

This creates two nodes, A and B, with A linking to B, and sets the fill color of A to green.
}
\examples{
value_tree(
  diagram =
    "graph LR;
  A(<B>Benefit-Risk Balance</B>)-->B(<B>Benefits</B>)
  B-->C(<B>Primary Efficacy</B>)
  B-->D(<B>Secondary Efficacy</B>)
  B-->E(<B>Quality of life</B>)
  C-->F(<B>\% Success</B>)
  D-->G(<B>Mean change</B>)
  E-->H(<B>Mean change</B>)
  A-->I(<B>Risks</B>)
  I-->J(<B>Recurring AE</B>)
  I-->K(<B>Rare SAE</B>)
  I-->L(<B>Liver Toxicity</B>)
  J-->M(<B>Event rate</B>)
  K-->N(<B>\% Event</B>)
  L-->O(<B>\% Event</B>)
  style A fill:#7ABD7E
  style B fill:#7ABD7E
  style I fill:#7ABD7E
  style C fill:#FFE733
  style D fill:#FFE733
  style E fill:#FFE733
  style J fill:#FFE733
  style K fill:#FFE733
  style L fill:#C6C6C6
  style F fill: #FFAA1C
  style G fill: #FFAA1C
  style H fill: #FFAA1C
  style M fill: #FFAA1C
  style N fill: #FFAA1C
  style O fill: #C6C6C6
  "
)
}
